Perform a more robust comparison of columns and data types
